requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
python-multipart==0.0.9
pydantic==2.9.2
scikit-learn==1.5.1
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
requests==2.32.3
python-dotenv
google-generativeai